requirements.txt
streamlit
openai-whisper
torch
librosa
jiwer
soundfile
numpy
